arrival_and_departure_tile.dart
Added on press -> gets all the stops for a route if clicked on from the bottom screen. Updates in a new provider called RouteProvider.dart

Added some todos to the check_auth.dart

Added new RouteProvider.dart to main.dart

Added instance of RouteProvider and made google map widget a consumer of the provider. Polylines now updates depending on which route you click.

Added getStopsForRoute function to state_info.dart so all stops for a given route could be retrieved.